+++ CheckAhead/src/main/java/com/skilldistillery/checkahead/services/ReviewRatingServiceImpl.java
@@ -13,6 +13,7 @@ import com.skilldistillery.checkahead.entities.ReviewRatingId;
 import com.skilldistillery.checkahead.repositories.RatingRepository;
 import com.skilldistillery.checkahead.repositories.ReviewRatingRepository;
 import com.skilldistillery.checkahead.repositories.ReviewRepository;
+import com.skilldistillery.checkahead.repositories.UserRepository;
 
 //abbreviated ReviewRating to rr or RR
 @Service
@@ -24,6 +25,8 @@ public class ReviewRatingServiceImpl implements ReviewRatingService {
 	ReviewRepository reviewRepo;
 	@Autowired
 	RatingRepository ratingRepo;
+	@Autowired
+	UserRepository userRepo;
 	
 	@Override
 	public List<ReviewRating> findAllRRs() {
@@ -31,46 +34,56 @@ public class ReviewRatingServiceImpl implements ReviewRatingService {
 	}
 	
 	@Override
-	public ReviewRating createRR(int reviewId, int ratingId, ReviewRating rr) {
-		Optional<Review> review = reviewRepo.findById(reviewId);
-		if (review.isPresent()) {
-			rr.setReview(review.get());
-		}
-		else {
+	public ReviewRating createRR(int reviewId, int ratingId, ReviewRating rr, String username) {
+//		User user = userRepo.findByUsername(username);
+		try {
+			Optional<Review> review = reviewRepo.findById(reviewId);
+			if (review.isPresent()) {
+				rr.setReview(review.get());
+			}
+			else {
+				return null;
+			}
+			Optional<Rating> rating = ratingRepo.findById(ratingId);
+			if (rating.isPresent()) {
+				rr.setRating(rating.get());
+			}
+			else {
+				return null;
+			}
+			if(rr.getReview().getUser().getUsername().equals(username)){
+				ReviewRating newrr = rrRepo.saveAndFlush(rr);
+				if (newrr != null) {
+					return newrr;
+				}
+			}
 			return null;
-		}
-		Optional<Rating> rating = ratingRepo.findById(ratingId);
-		if (rating.isPresent()) {
-			rr.setRating(rating.get());
-		}
-		else {
+		} catch (Exception e) {
+			e.printStackTrace();
 			return null;
-		}
-		ReviewRating newrr = rrRepo.saveAndFlush(rr);
-		if (newrr != null) {
-			return newrr;
-		}
-			return null;			
+		}			
 	}
 	
 	@Override
-	public ReviewRating updateRR(int id, ReviewRating rr) {
+	public ReviewRating updateRR(int id, ReviewRating rr, String username) {
 		Optional<ReviewRating> oldRR = rrRepo.findById(rr.getId());
 		ReviewRating managedRR = null;
 		if (oldRR.isPresent()) {
 			managedRR = oldRR.get();
 			managedRR.setRatingValue(rr.getRatingValue());
-			return rrRepo.saveAndFlush(managedRR);			
+			if(managedRR.getReview().getUser().getUsername().equals(username)){
+				return rrRepo.saveAndFlush(managedRR);							
+			}
 		}
 		return null;
 	}
 	
 	@Override
-	public boolean deleteRR(int reviewId, int ratingId) {
+	public boolean deleteRR(int reviewId, int ratingId, String username) {
 		boolean answer = false;
 		ReviewRatingId rrId = new ReviewRatingId(reviewId,ratingId);
 		Optional<ReviewRating> rr = rrRepo.findById(rrId);
-		if (rr.isPresent()) {
+		if (rr.isPresent() && rr.get().getReview().getUser().getUsername().equals(username)) {
 			rrRepo.deleteById(rrId);
 			answer = true;
 		}
