+++ CheckAhead/src/main/java/com/skilldistillery/checkahead/services/ReviewServiceImpl.java
@@ -3,8 +3,6 @@ package com.skilldistillery.checkahead.services;
 import java.util.List;
 import java.util.Optional;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +12,6 @@ import com.skilldistillery.checkahead.entities.Location;
 import com.skilldistillery.checkahead.entities.Review;
 import com.skilldistillery.checkahead.entities.User;
 import com.skilldistillery.checkahead.repositories.ReviewRepository;
-import com.skilldistillery.checkahead.repositories.UserRepository;
 
 @Transactional
 @Service
@@ -24,8 +21,8 @@ public class ReviewServiceImpl implements ReviewService {
 	@Autowired
 	private ReviewRepository reviewRepo;
 
-	@Autowired
-	private UserRepository userRepo;
+//	@Autowired
+//	private UserRepository userRepo;
 
 	@Override
 	public List<Review> findAllReviews() {
@@ -43,12 +40,14 @@ public class ReviewServiceImpl implements ReviewService {
 	}
 
 	@Override
-	public Review createReview(User user, Review review, Location location) {
+	public Review createReview(User user, Review review, Location location, String username) {
 //		User user = userRepo.findByUsername(username);
 		if (user != null) {
 			review.setUser(user);
 			review.setLocation(location);
-			reviewRepo.saveAndFlush(review);
+			if (review.getUser().getUsername().equals(username)) {
+				reviewRepo.saveAndFlush(review);
+			}
 		} else {
 			review = null;
 		}
@@ -63,7 +62,7 @@ public class ReviewServiceImpl implements ReviewService {
 
 	
 	@Override
-	public Review updateReview(Integer reviewId, Review review) {
+	public Review updateReview(Integer reviewId, Review review, String username) {
 		Optional<Review> optReview = reviewRepo.findById(reviewId);
 		Review managedReview = null;
 		if (optReview.isPresent()) {
@@ -72,16 +71,18 @@ public class ReviewServiceImpl implements ReviewService {
 			managedReview.setContent(review.getContent());
 			managedReview.setUser(review.getUser());
 			managedReview.setLocation(review.getLocation());
-			return reviewRepo.saveAndFlush(managedReview);			
+			if (managedReview.getUser().getUsername().equals(username)) {
+				return reviewRepo.saveAndFlush(managedReview);
+			}			
 		}
 		return null;
 	}
 
 	@Override
-	public boolean deleteReview(Integer reviewId) {
+	public boolean deleteReview(Integer reviewId, String username) {
 		boolean answer = false;
 		Optional<Review> optReview = reviewRepo.findById(reviewId);
-		if (optReview.isPresent()) {
+		if (optReview.isPresent() && optReview.get().getUser().getUsername().equals(username)) {
 			reviewRepo.deleteById(reviewId);
 			answer = true;
 		}
